+++ CHANGELOG.md
@@ -74,7 +74,9 @@
 - #952 Dummy PR
 - #957 Fixed issues caused by changes to timespamp in cudf
 - #962 Use new rmm API instead of get_device_resource() and set_device_resource() functions
+- #965 Handle exceptions from pool_threads
 - #963 Set log_level when using LOGGING_LEVEL param
+
 
 # BlazingSQL 0.14.0 (June 9, 2020)